+++ app/routes/contact.tsx
@@ -18,30 +18,18 @@ export default function Contact() {
         <div className="bg-white dark:bg-gray-800 rounded-lg shadow-md p-8">
           <div className="prose dark:prose-invert max-w-none">
             <p className="text-gray-700 dark:text-gray-300 mb-4">
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed do
-              eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut
-              enim ad minim veniam, quis nostrud exercitation ullamco laboris
-              nisi ut aliquip ex ea commodo consequat.
+              Github link:{' '}
+              <a href="https://github.com/bela-varga/ajc-lib-search">
+                https://github.com/bela-varga/ajc-lib-search
+              </a>{' '}
+              (jelenleg privát repo)
             </p>
 
             <p className="text-gray-700 dark:text-gray-300 mb-4">
-              Duis aute irure dolor in reprehenderit in voluptate velit esse
-              cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat
-              cupidatat non proident, sunt in culpa qui officia deserunt mollit
-              anim id est laborum.
-            </p>
-
-            <p className="text-gray-700 dark:text-gray-300 mb-4">
-              Sed ut perspiciatis unde omnis iste natus error sit voluptatem
-              accusantium doloremque laudantium, totam rem aperiam, eaque ipsa
-              quae ab illo inventore veritatis et quasi architecto beatae vitae
-              dicta sunt explicabo.
-            </p>
-
-            <p className="text-gray-700 dark:text-gray-300">
-              Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit
-              aut fugit, sed quia consequuntur magni dolores eos qui ratione
-              voluptatem sequi nesciunt.
+              Email:{' '}
+              <a href="mailto:bela.varga.work@gmail.com">
+                bela.varga.work@gmail.com
+              </a>
             </p>
           </div>
         </div>